scripts/requirements.txt
tensorflow>=2.20.0
numpy>=1.21.0
Pillow>=8.3.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
